Fixed unordered_map resize and exception
unordered_map resize now works.

Instead of having undefined behavior when an exception is thrown outside
of a try-catch, it would now call terminate. But keep in mind, this fix
has the cause of not allowing the user to return from a function within
a try catch block as that would exit the single-iteration for loop which
allows for popping the try frame after the try block is executed.